pdfparser/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
easyocr==1.7.1
environs==11.0.0
filelock==3.15.4
fsspec==2024.6.1
idna==3.7
imageio==2.35.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==2.1.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pyclipper==1.3.0.post5
python-bidi==0.6.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
setuptools==73.0.1
shapely==2.0.6
sqlparse==0.5.1
sympy==1.13.2
tifffile==2024.8.10
torch==2.4.0
torchvision==0.19.0
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
celery==5.4.0
redis==4.0.0
